bug fix with copulating child (didnt get correct connections) added unit test to cover that

+++ src/main/java/com/hkamran/ai/NeatFitness.java
@@ -2,6 +2,6 @@ package com.hkamran.ai;
 
 public interface NeatFitness {
 
-	public long calculate(Network network);
+	public double calculate(Network network);
 	
 }
